moved login component to components folder
